Renamed traits.
