+++ src/db-connection/src/Traits/HasRepository.php
@@ -11,36 +11,15 @@ declare(strict_types=1);
  */
 namespace Hyperf\DbConnection\Traits;
 
-use Hyperf\Database\ConnectionInterface;
-use Hyperf\Database\ConnectionResolverInterface;
 use Hyperf\Database\Model\Model;
-use Hyperf\Utils\ApplicationContext;
-use Psr\Container\ContainerInterface;
-use Psr\EventDispatcher\EventDispatcherInterface;
 use RuntimeException;
 
-/** @mixin Model */
-trait Repository
+trait HasRepository
 {
     /**
      * @var string the full namespace of repository class
      */
     protected $repository;
-
-    /**
-     * Get the database connection for the model.
-     */
-    public function getConnection(): ConnectionInterface
-    {
-        $connectionName = $this->getConnectionName();
-        $resolver = $this->getContainer()->get(ConnectionResolverInterface::class);
-        return $resolver->connection($connectionName);
-    }
-
-    public function getEventDispatcher(): ?EventDispatcherInterface
-    {
-        return $this->getContainer()->get(EventDispatcherInterface::class);
-    }
 
     /**
      * @throws RuntimeException when the model does not define the repository class
@@ -51,10 +30,5 @@ trait Repository
             throw new RuntimeException(sprintf('Cannot detect the repository of %s', static::class));
         }
         return $this->getContainer()->get($this->repository);
-    }
-
-    protected function getContainer(): ContainerInterface
-    {
-        return ApplicationContext::getContainer();
     }
 }
